added CRUD for price and discount

+++ src/main/java/kg/mega/naTV/entities/dto/ChannelsDto.java
@@ -1,6 +1,6 @@
 package kg.mega.naTV.entities.dto;
 
-import kg.mega.naTV.entities.Discounts;
+import kg.mega.naTV.controller.PriceController;
 import kg.mega.naTV.entities.Price;
 import lombok.Data;
 
@@ -9,7 +9,8 @@ public class ChannelsDto {
     private Long id;
     private String channelName;
     private String logo;
-    private Integer pricePerLetter;
+    private Double pricePerLetter;
     private DiscountDto discountDto;
+
     
 }
